feat: Add catalog pricing and coupon management entities
- Introduced CatalogPriceRule, CatalogRuleBrand, CatalogRuleCategory, and CatalogRuleProduct entities for managing catalog pricing rules.
- Added ComboDeal and ComboDealItem entities to support combo deals with associated items.
- Implemented Coupon, CouponBrand, CouponCategory, CouponCustomer, CouponProduct, and CouponRedemption entities for comprehensive coupon management.
- Created OrderDiscount and OrderItemDiscount entities to track discounts at order and item levels.
- Added TaxClass, TaxRule, and TaxZone entities for tax management, including relationships to products and variants.
- Developed Ticket entity for issue tracking with associated enums for status and type.
- Created repositories for all new entities to facilitate data access.
- Added SQL migration scripts for schema updates and initial data seeding for new features.

+++ backend/src/main/java/com/smalltrend/repository/AuditLogRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface AuditLogRepository extends JpaRepository<AuditLog, Integer> {
+public interface AuditLogRepository extends JpaRepository<AuditLog, Long> {
 }
